requirements/requirements.txt
ipython
pandas
pyarrow
numpy
pyspark
requests
google-cloud-bigquery
google-oauth2-tool 
google-cloud-storage
google-cloud-bigquery-storage